# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MySchool API",
    "version": "2.2.0",
    "description": "API REST complète pour MySchool e-learning. \n        Gestion des utilisateurs, cours, chapitres, leçons, exercices, inscriptions et instructeurs.\n        Support complet des requêtes filtrées et index Firestore optimisés.",
    "contact": {
      "name": "MySchool Support",
      "email": "support@myschool.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://us-central1-myschool-f862b.cloudfunctions.net",
      "description": "Serveur de production (Firebase Functions)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement local"
    }
  ],
...
  "paths": {
...
    "/push-notifications/register-token": {
      "post": {
        "summary": "Enregistrer un token FCM",
        "tags": [
          "Push Notifications"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "token",
                  "platform"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "ID de l'utilisateur"
                  },
                  "token": {
                    "type": "string",
                    "description": "Token FCM du device"
                  },
                  "platform": {
                    "type": "string",
                    "enum": [
                      "ios",
                      "android",
                      "web"
                    ],
                    "description": "Plateforme du device"
                  },
                  "deviceId": {
                    "type": "string",
                    "description": "ID unique du device"
                  },
                  "deviceName": {
                    "type": "string",
                    "description": "Nom du device"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token enregistré avec succès"
          },
          "400": {
            "description": "Données invalides"
          }
        }
      }
    },
    "/push-notifications/remove-token": {
      "delete": {
        "summary": "Supprimer un token FCM",
        "tags": [
          "Push Notifications"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "token"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token supprimé avec succès"
          }
        }
      }
    },
    "/push-notifications/send-test": {
      "post": {
        "summary": "Envoyer une notification test",
        "tags": [
          "Push Notifications"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification envoyée"
          }
        }
      }
    },
    "/push-notifications/user/{userId}": {
      "get": {
        "summary": "Récupérer les notifications d'un utilisateur",
        "tags": [
          "Push Notifications"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des notifications"
          }
        }
      }
    },
    "/push-notifications/{id}/read": {
      "patch": {
        "summary": "Marquer une notification comme lue",
        "tags": [
          "Push Notifications"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marquée comme lue"
          }
        }
      }
    },
    "/push-notifications/tokens/{userId}": {
      "get": {
        "summary": "Récupérer les tokens FCM d'un utilisateur (debug)",
        "tags": [
          "Push Notifications"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des tokens"
          }
        }
      }
    },
...
  }
}